+++ app/renderer/actions.js
@@ -563,7 +563,12 @@ const toCommandMode = () => {
     setMode("command")
 }
 
-const stopLoadingPage = () => currentPage()?.stop()
+const stopLoadingPage = args => {
+    if (args?.key === "<C-c>") {
+        notify("This mapping is deprecated, replaced with <Esc>", "warn")
+    }
+    currentPage()?.stop()
+}
 
 const moveTabForward = () => {
     const {"moveTabForward": move} = require("./tabs")